% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_carbon}
\alias{calculate_carbon}
\title{Calcul du carbone à partir de la biomasse}
\usage{
calculate_carbon(df)
}
\arguments{
\item{df}{Le data.frame contenant la colonne Biomass_Total}
}
\value{
Le data.frame enrichi avec la colonne Carbon_Total
}
\description{
Calcule le carbone total comme 50% de la biomasse totale.
}
