% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_prediction_interval}
\alias{calculate_prediction_interval}
\title{Calcul de la variance individuelle et des intervalles de prédiction}
\usage{
calculate_prediction_interval(
  x,
  eqs_volume,
  equation_id = 1,
  confidence_level = 0.95
)
}
\arguments{
\item{equation_id}{Identifiant de l'équation utilisée}

\item{df}{Le data.frame contenant les volumes calculés}

\item{volume_type}{Le nom de la colonne de volume (ex: "V22")}
}
\value{
Le data.frame enrichi avec les colonnes Relative_Interval_Width et Interval_Interpretation
}
\description{
Calcule la largeur relative des intervalles de prédiction pour chaque arbre.
}
