% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomasse.R
\name{calculer_biomasse}
\alias{calculer_biomasse}
\title{Calcule la biomasse aérienne et le carbone associé à partir des volumes et coefficients}
\usage{
calculer_biomasse(
  donnees_arbres,
  essence = NULL,
  colonne_volume = NULL,
  afficher_warnings = TRUE
)
}
\arguments{
\item{donnees_arbres}{Un data frame contenant les colonnes 'Essence', 'VC22' ou 'VC22B'}

\item{essence}{Essence ciblée (ou NULL pour utiliser celle de chaque ligne)}

\item{colonne_volume}{Nom de la colonne de volume à utiliser (NULL utilise celle spécifiée dans les équations internes)}

\item{afficher_warnings}{Booléen indiquant si les warnings doivent être affichés (TRUE) ou stockés (FALSE)}
}
\value{
Le data frame d'entrée avec les colonnes supplémentaires 'Biomasse', 'Carbone' et 'Messages' si afficher_warnings=FALSE
}
\description{
Calcule la biomasse aérienne et le carbone associé à partir des volumes et coefficients
}
\examples{
donnees <- data.frame(
  Essence = c("Chene", "Hetre", "Pin"),
  VC22 = c(120, 150, 100)
)

calculer_biomasse(donnees)
}
