% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbofor.R
\name{carbofor}
\alias{carbofor}
\title{Calculate Tree Volumes Using Allometric Equations}
\usage{
carbofor(x, ...)
}
\arguments{
\item{x}{An object (usually a data.frame) containing tree data.}

\item{...}{Additional arguments passed to methods.}
}
\description{
This is a generic function for calculating tree volumes using species-specific allometric equations.
See method documentation for details.
}
