% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{diameter_conversions}
\alias{diameter_conversions}
\title{Conversion diamètre <-> circonférence}
\usage{
diameter_conversions(x, C130, C150, D130, D150, HTOT = "HTOT", HDOM = "HDOM")
}
\description{
Conversion diamètre <-> circonférence
}
