% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{double}
\alias{double}
\title{Double}
\usage{
double(number)
}
\arguments{
\item{number}{un vecteur numérique}
}
\value{
un vecteur de meme longueur que le number
}
\description{
Double la valeur d'un nombre
}
\details{
Calcule le double de la valeur d'un nombre
}
\examples{
double(2)
double(1:4)
}
