% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Prétraitement des données pour le calcul de volume}
\usage{
preprocess_data(
  x,
  specimens = NULL,
  C130 = "C130",
  C150 = "C150",
  D130 = "D130",
  D150 = "D150",
  HTOT = "HTOT",
  HDOM = "HDOM",
  ...
)
}
\description{
Prétraitement des données pour le calcul de volume
}
