% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{update_flags}
\alias{update_flags}
\title{Update existence flags for key column names}
\usage{
update_flags(x, flags, C130, C150, D130, D150, HTOT, HDOM)
}
\arguments{
\item{x}{A data.frame containing tree data.}

\item{flags}{A named list of logicals tracking existence of key variables.}

\item{C130, }{C150, D130, D150, HTOT, HDOM Column names to check in the data.frame.}
}
\value{
An updated list of logical flags indicating column presence.
}
\description{
Checks whether key variables (diameters, circumferences, heights) exist in the input data
and updates a logical flag list accordingly.
}
