% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Valide les paramètres d'entrée pour la fonction carbofor}
\usage{
validate_parameters(
  x,
  volume_type = "V22",
  equation_id = 1,
  source = "Dagnellie",
  specimens = NULL,
  C130 = "C130",
  C150 = "C150",
  D130 = "D130",
  D150 = "D150",
  HTOT = "HTOT",
  HDOM = "HDOM"
)
}
\arguments{
\item{x}{Le data.frame d'entrée}

\item{volume_type}{Type de volume à calculer}

\item{equation_id}{Identifiant de l'équation}

\item{source}{Source des équations ("Dagnellie", "Aflan", "Vallet")}

\item{specimens}{Nom de la colonne d'identification des espèces}

\item{C130, }{C150, D130, D150, HTOT, HDOM Noms des colonnes dendrométriques}
}
\description{
Valide les paramètres d'entrée pour la fonction carbofor
}
