% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_bark_thickness}
\alias{calculate_bark_thickness}
\title{Calcul de l'épaisseur et du volume d'écorce}
\usage{
calculate_bark_thickness(df, total_volume_col = "V22")
}
\arguments{
\item{df}{Le data.frame contenant les données d'arbres}

\item{total_volume_col}{Nom de la colonne contenant le volume total (ex: "V22")}
}
\value{
Le data.frame enrichi avec les colonnes E, Bark_Volume et Wood_Volume
}
\description{
Utilise les équations spécifiques (type E) pour estimer l'épaisseur d'écorce, puis calcule le volume d'écorce et le volume de bois net.
}
