% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{calculate_basal_areas}
\alias{calculate_basal_areas}
\title{Calcul des surfaces basales G130 et G150}
\usage{
calculate_basal_areas(df, C130, C150)
}
\description{
Calcul des surfaces basales G130 et G150
}
