% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_biomass}
\alias{calculate_biomass}
\title{Calcul de la biomasse à partir des équations spécifiques}
\usage{
calculate_biomass(df)
}
\arguments{
\item{df}{Le data.frame contenant les données d'arbres}
}
\value{
Le data.frame enrichi avec les colonnes de biomasse et carbone
}
\description{
Calcule la biomasse pour chaque arbre en appliquant toutes les équations disponibles pour son espèce.
}
