% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_carbon}
\alias{calculate_carbon}
\title{Carbon calculation from biomass}
\usage{
calculate_carbon(x)
}
\arguments{
\item{x}{A data.frame containing a Biomass_Total column}
}
\value{
The same data.frame enriched with the Carbon Total column
}
\description{
Calculates total carbon as 50% of total biomass.
}
\examples{
x <- data.frame(Biomass_Total = c(100, 120, 80))
x <- calculate_carbon(x)
}
