% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_volume}
\alias{calculate_volume}
\title{Calcul du volume à partir des équations allométriques}
\usage{
calculate_volume(
  df,
  volume_type = "V22",
  equation_id = 1,
  source = "Dagnellie"
)
}
\arguments{
\item{df}{Le data.frame contenant les données d'arbres}

\item{volume_type}{Type de volume à calculer (ex: "V22", "V22B", "V22_HA", "E")}

\item{equation_id}{Identifiant de l'équation à utiliser}

\item{source}{Source des équations (ex: "Dagnellie")}
}
\value{
Le data.frame enrichi avec les colonnes de volume, validité et équation utilisée
}
\description{
Applique les équations allométriques spécifiques à chaque espèce pour calculer le volume.
}
