% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{calculer_volumes}
\alias{calculer_volumes}
\title{Calcule les volumes forestiers selon differentes methodes}
\usage{
calculer_volumes(
  df,
  type_volume = "VC22",
  essence = NULL,
  equations_df = equations,
  id_equation = 1,
  coefs_conversion = NULL,
  remove_na = FALSE
)
}
\arguments{
\item{df}{Un data frame contenant les donnees des arbres avec au moins une colonne d'identification
d'essence ("Essence", "Code", ou "Abr")}

\item{type_volume}{Type de volume a calculer (ex: "VC22", "VC22B", "E", "VC22_HA")}

\item{essence}{Essence d'arbre specifique. Si NULL (par defaut), l'essence est lue ligne par ligne dans `df`.}

\item{equations_df}{Le data frame contenant les equations allometriques a appliquer}

\item{id_equation}{Indice de l'equation a utiliser si plusieurs sont disponibles (par defaut = 1)}

\item{coefs_conversion}{(Optionnel) Data frame pour convertir C150 en C130 (doit contenir "Essence" et "Coef_C150_C130")}

\item{remove_na}{(Optionnel) Si TRUE, supprime les lignes où aucun volume n'a pu etre calcule (par defaut = FALSE)}
}
\value{
Un data frame avec les colonnes d'origine augmentees de :
\itemize{
  \item `Volume` : volume calcule selon l'equation choisie
  \item `Equation_Utilisee` : description de l'equation utilisee
}
}
\description{
Cette fonction applique des equations allometriques pour estimer le volume
des arbres a partir de differentes variables dendrometriques.
}
\details{
Le champ `A0` dans les equations allometriques indique le type de formule utilisee :
\itemize{
  \item{1, 2, 3, 5 : Formule lineaire du type }{ V = b0 + b1*X1 + b2*X2 + ... }
  \item{4 : Formule logarithmique du type }{ V = 10^(b0 + b1 * log10(X1)) }
}
}
