% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{calculer_volumes}
\alias{calculer_volumes}
\title{Calcule les volumes forestiers selon différentes méthodes}
\usage{
calculer_volumes(
  donnees_arbres,
  type_volume = "VC22",
  essence = NULL,
  equations_df = equations_allometriques,
  id_equation = 1,
  coefs_conversion = NULL
)
}
\arguments{
\item{donnees_arbres}{Un data frame contenant les données des arbres}

\item{type_volume}{Type de volume à calculer (ex: "VC22", "VC22B", "Ecorce")}

\item{essence}{Essence d'arbre spécifique (ou NULL pour utiliser celle des données)}

\item{equations_df}{Le data frame contenant les équations allométriques}

\item{id_equation}{Indice de l'équation à utiliser si plusieurs sont disponibles (par défaut = 1)}

\item{coefs_conversion}{(Optionnel) Data frame de conversion de C150 vers C130}
}
\value{
Un data frame avec les données d'entrée plus une colonne de volume calculé
}
\description{
Calcule les volumes forestiers selon différentes méthodes
}
