% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{convert_circumference}
\alias{convert_circumference}
\title{Conversion générique entre C130 et C150}
\usage{
convert_circumference(df, C130, C150)
}
\description{
Conversion générique entre C130 et C150
}
