% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{convert_circumference}
\alias{convert_circumference}
\title{Conversion générique entre C130 et C150}
\usage{
convert_circumference(
  x,
  D150 = "D150",
  D130 = "D130",
  C150 = "C150",
  C130 = "C130",
  HTOT = "HTOT",
  HDOM = "HDOM"
)
}
\description{
Conversion générique entre C130 et C150
}
