% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liste.R
\name{liste_eq}
\alias{liste_eq}
\title{Liste les équations disponibles pour une essence et un type de volume donné}
\usage{
liste_eq(essence = NULL, type_volume = NULL)
}
\arguments{
\item{essence}{(Optionnel) Le nom de l'essence à filtrer (ex : `"Bouleau"`). Si NULL, toutes les essences sont affichées.}

\item{type_volume}{(Optionnel) Le type de volume à filtrer (ex : `"VC22"`, `"E"`, `"VC22B"`). Si NULL, tous les types de volume sont affichés.}
}
\value{
Un data frame avec un identifiant (`ID`) et les colonnes utiles pour visualiser les équations :
`Y`, `Essences`, `X1:X5`, `b0:b5`.
}
\description{
Liste les équations disponibles pour une essence et un type de volume donné
}
